Add/remove author information and hard-wrap
In most of my articles, I had not mentioned who the author is
and I referred to myself using words like I, my, mine, etc.
If these articles are supposed to be used for a long time and
edited by many people, this is not desirable.
From some articles, I removed references because others who
reviewed the articles and provided valuable inputs also
agreed to the views expressed in the articles.
For articles which are specifically about my way of doing things,
I have clearly mentioned myself as the author (like in django.md).

Also hard-wrapped lines in all files.